pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "second-brain-telegram-aws"
version = "1.0.0"
description = "Personal second brain system using Telegram and AWS Lambda"
authors = [
    {name = "Second Brain", email = "noreply@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["telegram", "aws", "lambda", "second-brain", "ai", "productivity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Office/Business :: Groupware",
]

dependencies = [
    "boto3>=1.34.0",
    "requests>=2.31.0", 
    "anthropic>=0.21.0",
    "openai>=1.10.0",
    "InquirerPy>=0.3.4",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
]

[project.scripts]
setup-webhook = "scripts.setup_webhook:main"
webhook-setup = "scripts.setup_webhook:main"
telegram-webhook = "scripts.setup_webhook:main"
secrets = "scripts.secrets:main"
cdk = "cdk_app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["processor*", "digest*", "scripts*"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]